+++ addOns/pscanrules/src/test/java/org/zaproxy/zap/extension/pscanrules/InsecureJSFViewStatePassiveScannerUnitTest.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.extension.pscanrulesBeta;
+package org.zaproxy.zap.extension.pscanrules;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
@@ -33,7 +33,7 @@ import org.parosproxy.paros.network.HttpMessage;
 public class InsecureJSFViewStatePassiveScannerUnitTest extends
         PassiveScannerTest<InsecureJSFViewStatePassiveScanner> {
 
-    private static final String BASE_RESOURCE_KEY = "pscanbeta.insecurejsfviewstate.";
+    private static final String BASE_RESOURCE_KEY = "pscanrules.insecurejsfviewstate.";
     private static final String INSECURE_JSF = BASE_RESOURCE_KEY + "name";
 
     @Override