+++ addOns/pscanrulesBeta/CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix typo in scanner name.
 - Tweak alert Authentication Credentials Captured to use Description field instead of Attack.
 - Remove Charset Mismatch Scanner (promoted to release Issue 4460).
+- Remove ViewState Scanner (promoted to release Issue 4453).
+- Remove Insecure JSF ViewState Scanner (promoted to release Issue 4455).
+- Remove Insecure Authentication Scanner (promoted to release Issue 4456).
+- Remove Information Disclosure Debug Errors Scanner (promoted to release Issue 4457).
+- Remove CSRF Countermeasures Scanner (promoted to release Issue 4458).
+- Remove Cookie Loosely Scoped Scanner (promoted to release Issue 4459).
 
 ## 18 - 2018-01-19